fix: MAC softmax float32 cast + paper table updates + phenoflux_crispr config
- mac.py: add .float() cast before softmax in MAC attention, matching
  UNet's existing attention pattern. Fixes potential NaN gradients under
  mixed-precision training (float16 softmax overflow).
- PAPER_TABLES.md: fill MoA column with 5K confirmed results
  (CellFlux=76.66%, StarGAN=46.3%, IMPA=48.5%, PhenDiff=60.69%).
- RESULTS.md: update CellFlux-Style Diet Table with 5K numbers.
- model_configs.py: add phenoflux_crispr config (condition_dim=204,
  MAC+CCM modules for CRISPR gene-conditioned transport).
- train.py: add phenoflux_crispr to recognized dataset list.

+++ src/morphoflux/engine/models/mac.py
@@ -152,7 +152,7 @@ class MACAttentionBlock(nn.Module):
         # Scaled dot-product attention
         scale = self.head_dim ** -0.5
         attn = torch.matmul(q, k.transpose(-2, -1)) * scale      # (B, heads, HW, N_ctx)
-        attn = torch.softmax(attn, dim=-1)
+        attn = torch.softmax(attn.float(), dim=-1).type(attn.dtype)
         out = torch.matmul(attn, v)                               # (B, heads, HW, head_dim)
 
         # Merge heads -> project -> residual
